Make taxonomy selection more agnostic (#9037)
